refactor: move timeline files in component folder
